add gcc 4.4 support for Loongson2
because only gcc 4.4 have loongson-specific support, so, we need to
choose the suitable -march argument for gcc <= 4.3 and gcc >= 4.4, and
we also need to consider use -march=loongson2e and -march=loongson2f for
loongson2e and loongson2f respectively. this is handled by adding two
new kernel options: CPU_LOONGSON2E and CPU_LOONGSON2F(thanks for the
solutin provided by ZhangLe).

I have tested it on FuLoong(2f) in 32bit and 64bit with gcc-4.4 and
gcc-4.3. so, basically, it works.

thanks goes to Arnaud Patard <apatard@mandriva.com> for suggestion the
usage of cc-options(Documentation/kbuild/makefiles.txt).

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>