only enable -mfix-ls2f-kernel for loongson2f
In function as-option, the parameter is passed to CC, without "-Wa," prefix. So
the parameter should have "-Wa," prefix as shown in script/Kbuild.include. Also
because ,(comma) is as-option's parameter delimeter, the comma in "-Wa," should
be replaced by $(comma).

Documentation/kbuild/makefiles.txt:

as-option = $(call try-run,\
        $(CC) $(KBUILD_CFLAGS) $(1) -c -xassembler /dev/null -o "$$TMP",$(1),$(2))

and currently, there is no proof of need to use -mfix-ls2f-kernel for
non-loongson2f, so just enable it for loongson2f.

Signed-off-by: Zhang Le <r0bertz@gentoo.org>
Signed-off-by: Wu Zhangjin <wuzj@lemote.com>