only enable -mfix-ls2f-kernel for loongson2f-based machines
In function as-option, the parameter is passed to CC, without "-Wa," prefix. So
the parameter should have "-Wa," prefix as shown in script/Kbuild.include. Also
because ,(comma) is as-option's parameter delimeter, the comma in "-Wa," should
be replaced by $(comma).

Documentation/kbuild/makefiles.txt:

as-option = $(call try-run,\
        $(CC) $(KBUILD_CFLAGS) $(1) -c -xassembler /dev/null -o "$$TMP",$(1),$(2))

in reality, this option is only needed for loongson2f-based machines,
   this is not a bug of cpu itself, but a potential cooperation problem
   between cpu and cs5536 bridge.

Signed-off-by: Zhang Le <r0bertz@gentoo.org>
Signed-off-by: Wu Zhangjin <wuzj@lemote.com>

Conflicts:

	arch/mips/Makefile

	as the manpage of gcc says, if not indicate -mtune, which will
	optimize the arch indicated in -march, so there is no need to
	indicate -mtune and -march to the same arch. just remove the -mtune
	option.

Signed-off-by: Wu Zhangjin <wuzj@lemote.com>