[loongson] add basic STR/Standby support for loongson-based machines
This patch add the basic STR/Standby support for loongson-based
machines, for lacking for hardware support, we can only support Standby
here.

This patch benefit from the wait mode of loongson processor, when we set
the frequency of it to ZERO, it will enter into a wait mode, the power
consuption of it will become 0~1W, and if there is an external interrupt
invoked to it, it will be waked up.

Herein, to avoid modifying the hardware, we use the keyboard as the
interrupt source, so after suspending the keyboard device, we will
reenable the keyboard port and enable the keyboard irq in i8259A, then
the keyboard interrupt can be sent to loongson.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>