[loongson] compressed kernel support
This patch add compressed kernel support for loongson-based machines,
which will help to generate smaller kernel image. by default, after
compiling, there will be a compressed kernel image file stored in
arch/mips/loongson/image/bzImage, and there is a link of it in the top
directory named vmlinuz.

This patch is almost a self-contained implementation, include it's own
zlib, serial port debug interface. perhaps you ask why not use the
existing zlib and serial port impmentation in kernel, I should tell you
there is not good currently, because we can not access them in the
original vmlinux before decompressing it. and of course, there is an
possibility to share the exisiting zlib, serial port support, but which
is not easy to link them in, fore there are lots of dependences.

This implmentation is based on the zboot support in the linux_loongson
git repository on http://dev.lemote.com/

And this new implementation is almost ready to upstream for it's new
organization and cleanup. but some more things need to clean up.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>