[loongson] zImage: add whole new zImage support for loongson-based machines
This is a whole new zImage support for loongson-based machines, which is
based on the original compressed kernel support in the linux_loongson
git repository on http://dev.lemote.com, but this implementation not
only cleaned up the files, coding styles a lot of the original one, but
add the following features:

1. support 64bit

$ file arch/mips/loongson/image/zImage
arch/mips/loongson/image/zImage: ELF 64-bit LSB executable, MIPS, MIPS-III version 1 (SYSV), statically
linked, not stripped

2. support three different kernel compressing algorithm

inlucde GZIP, BZIP2 & LZMA, LZMA is the best one, but may need more time
to decompress, so, it's your choice to use which one.

and I think this is relly ready to the other mips platform, so, after a
few more test on the other mips board, I will try to push this patch to
the mainline linux kernel.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>