[loongson] merge fnkey into kernel source code: tools/yeeloong-2f/fnkey
fnkey is a tool for handling function keys(including LID, AC
plugin/out...) of YeeLoong laptop, it is originally written by Cui
Chunbo<cuicb@lemote.com> from Lemote. this version is based on his 1.0
version but does a lot of changes on the fnkey.c.

Although fnkey is not a kernel function but it depends on the kernel
interface: /proc/sci provided by the EC module. and Because the current
EC module implementation is not mature and need to tune a lot, so, I
think it's better to merge this tool to kernel source code and make it
available to the users of this git branch.

NOTE:
   you need to install this tool on your YeeLoong laptop. please get
   more information from the following documents.
Usage:
   tools/yeeloong-2f/fnkey/README
Install:
   tools/yeeloong-2f/fnkey/INSTALL

TODO:

Currently, fnkey works just like the addon-input of HAL does,
addon-input depends on the standard interfaces of input devices provided
in the /sys filesystem, but the EC module of YeeLoong does not provide
any of these interfaces, so we can not use addon-input directly.

and therefore, the not-long term plan is to provide standard interfaces
via change the EC module, and at last, discard this fnkey tool.

and please ignore the ERRORS in this tool reported by
./script/checkpatch.pl, because, the external xosd library(for drawing
fonts or pictures on screen) not follow the Coding Style of Linux. I
really do not want to touch it.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>