[loongson] add the backlight device as the cooling device for yeeloong2f
Because the fan can tuned by EC automatically, and the highest power
consuming of the loongson2f processor is only 4w, so, we only add the
backlight device as the cooling device. because it almost consume 1/3 of
the whole power.

And to reduce the power consuming, we initialize an appropriate
brightness to MAX_BRIGHTNESS/2.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>