[loongson] enable -mfix-ls2f-kernel for loongson2f
In function as-option, the parameter is passed to CC, without "-Wa,"
prefix. So the parameter should have "-Wa," prefix as shown in
script/Kbuild.include. Also because ,(comma) is as-option's parameter
delimeter, the comma in "-Wa," should be replaced by $(comma).

Documentation/kbuild/makefiles.txt:

as-option = $(call try-run,\
        $(CC) $(KBUILD_CFLAGS) $(1) -c -xassembler /dev/null -o "$$TMP",$(1),$(2))

and currently, there is no proof of need to use -mfix-ls2f-kernel for
non-loongson2f, so just enable it for loongson2f.

This option -mfix-ls2f-kernel is provided by the as patch from Lemote:
http://archlinux-for-loongson.googlegroups.com/web/binutils-2.19.1-loongson2f-3.patch

Signed-off-by: Zhang Le <r0bertz@gentoo.org>
Signed-off-by: Wu Zhangjin <wuzj@lemote.com>

and if the argument of the -tune is the same as -march, there is only a
need to use -march, so remove the -tune option here too.

Conflicts:

	arch/mips/Makefile

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>