[loongson] r8187b rfkill: convert gpio polling method to SCI interrupt method
In the old implementation, the r8187b wifi driver use gpio polling
method to trace the status of KEY_WLAN(Fn+F5), in reality, there is SCI
interrupt to handle this, so, we use the SCI interrupt method instead.
This will improve the performance a lot.

And at the same time, the r8187b_wifi_set() interface is implemented,
so, users can change the rfkill status via user-space application(i.e
NetworkManager, version >= 0.7.1 are recommended) or through the
following interface(need execute as root):

Disable it.

$ echo 0 > /sys/class/rfkill/rfkill0/state

Enable it.

$ echo 1 > /sys/class/rfkill/rfkill0/state

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>