[loongson] fix the sync problem of brightness tuning between EC and OS
When User press the Fn+Up/Down keys to tune the brightness, we will
report these key events to the user-space application via /sys
filesystem, and then the user-space application will try to draw the
progress bar and then call the set_brightness() function in Kernel, but
in reality, the EC have done this job, so, there will be a sync problem
between EC and Kernel.

to avoid touching the EC source code, we hack set_brightness() a little
to just ensure it not set the brightness when EC have tuned it. this
will avoid the sync problem between EC and OS when User press the keys
too fast. but this is really only a workaround, the best solution is
clarify the interfaces among EC, kernel and even the user-space
applications.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>