Extended scheduling parameters structure added.
An extended scheduling parameter structure, sched_param_ex, is defined in
this commit, as the starting point for supporting task models more
sophisticated than fixed-priority.

One that is both popular and (hopefully!) general enough is the so-called
sporadic task model, in which tasks' computation is divided into instances,
each one with:
 * a (maximum/typical) execution time,
 * a minimum interval between the activation of two consecutive instances,
 * a time instant by which the computation of the instance must be completed.

The new sched_param_ex reflects this model, and thus allows for better
specification of time sensitive workloads typical, for example, in real-time,
control and/or continuous media applications.

Signed-off-by: Raistlin <raistlin@linux.it>