SCHED_DEADLINE: core of the scheduling class.
This commit introduces a new scheduling policy (SCHED_DEADLINE), implemented
in a new scheduling class (sched_deadline.c).

As of now, it implements the popular Earliest Deadline First (EDF) real-time
scheduling algorithm.
It basically means each (instance of each) task has a deadline, indicating the
time instant by which its computation has to be completed. The scheduler
always picks the task with the earliest deadline as the next to be executed.

Some more logic is added in order to avoid tasks interfering between each
others, i.e., a deadline miss of task A should not affect the capability of
task B to meet its own deadline.

Open issues:
 - this implementation is ``fully partitioned'', which means each task has to
   be bound to one processor at any given time. Turning it into ``global
   scheduling'' (i.e., migrations are allowed) is work in progress;
 - proper dealing with critical sections/rt-mutexes is also missing, and
   is also work in progress.

Signed-off-by: Raistlin <raistlin@linux.it>
Signed-off-by: Michael <trimarchimichael@yahoo.it>