SCHED_DEADLINE: added sched_*_ex syscalls.
This commits adds the new syscalls needed to set/get the parameters for
SCHED_DEADLINE scheduling policy. As it can be expected, they all deal with
sched_param_ex.

The new syscalls are:
 * sched_setscheduler_ex,
 * sched_setparam_ex,
 * sched_getparam_ex.

They have been added to x86, x86-64 and ARM only for now, since these are
the only architectures we are able to test... But adding the bits needed for
supporting other archs is more than straightforward...

Signed-off-by: Raistlin <raistlin@linux.it>