Loongson: Disable PAGE_SIZE_4KB and FLATMEM
With PAGE_SIZE_4KB, the kernel will hang on:

Kernel panic - not syncing: Attempted to kill init!

The possible reason is the cache aliases problem. And also, With 16kb
page size, there is no cache aliases problem.

And with FLATMEM, The STD will fail, and there are also some other
problems when using NFS or CIFS(samba).

So, we disable them by default. and also, you are recommended to use
CONFIG_64BIT instread of the CONFIG_32BIT.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>