Loongson: YeeLoong: Cleanup the Fn+F5 action
In the old implementation, I have reported the KEY_WLAN event to the
user-space and provide the rfkill interface of the rtl8187b to the
user-space, so, the user-space applications can response the KEY_WLAN
and switch the rfkill status of the rtl8187b. but also, I have register
a EVENT_WLAN handler in the yeeloong_laptop module to switch the rfkill
status in the rfkill support of rtl8187b, this will duplicate the job
and make the whole thing weird.

Changes:

	- Remove the EVENT_WLAN handling in the kernel space, let it be
	the user-space. just report the KEY_WLAN and provide the related
	interfaces is enough.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>