cyclictest: ensure it work with glibc 2.7 in linux-MIPS
In glibc 2.7, there is no tid member of the struct sigevent for MIPS, to
ensure cyclictest work, we need to use the _attribute sub-member
instead. this _attribute member is not used by the other part of
cyclictest, so, it's safe to use it.

refer to: /usr/include/bits/siginfo.h

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>