SCHED_DEADLINE: temp fixup of "BUG: scheduling while atomic"
There is a bug in the kernel with sched_deadline on single core MIPS
platform(YeeLoong netbook), it is reproducible all the time.

BUG: scheduling while atomic: sirq-hrtimer/0/0x00000001/10, CPU#0
Call Trace:
[<ffffffff8020c2dc>] dump_stack+0x8/0x40
[<ffffffff8020ccd0>] __schedule+0x450/0x538
[<ffffffff8020d124>] schedule+0x24/0x60
[<ffffffff80261650>] hrtimer_wait_for_timer+0x70/0xc0
[<ffffffff802616c0>] hrtimer_cancel+0x20/0x50
[<ffffffff80235eac>] T.1236+0xf4/0x110
[<ffffffff8020caf0>] __schedule+0x270/0x538
[<ffffffff8020d0d0>] preempt_schedule+0xa8/0xd8
[<ffffffff80239908>] deadline_timer+0xe0/0x110
[<ffffffff80260744>] run_hrtimer_softirq+0xc4/0x208
[<ffffffff80247464>] ksoftirqd+0x15c/0x320
[<ffffffff8025c284>] kthread+0x94/0xa0
[<ffffffff80213384>] kernel_thread_helper+0x1c/0x28

After tracing it, found out that the kernel never return from the
atomic_spin_unlock() in deadline_timer() of kernel/sched_deadline.c and
make the whole system print the above information all the time.

The atomic_spin_unlock() did call dec_preempt_count() to restore the
preempt_count() to 0, but after calling preempt_schedule(), it printed
the above info and ...

And then I traced the __schedule() function and found out the
finish_task_switch()(called by context_switch and at last called by
__schedule()) tried to call hrtimer_cancel() to cancel
the hrtimer for the dead deadline task, but this time, the
deadline_timer is not finished yet, then we got deadlock:

deadline_timer(): hrtimer for deadline task
  -> atomic_spin_unlock:
     -> preempt_enable()
        -> preempt_check_resched()
           -> preempt_schedule()
              -> schedule()
                 -> __schedule
		    -> context_switch()
		       -> finish_task_switch(): T.1236
			  -> hrtimer_cancel()

So, the temp solution is not do preempt_schedule() in deadline_timer()
via calling atomic_spin_lock/unlock_bh() instead of the
atomic_spin_lock/unlock(), but this may reduce a potential scheduling
site, another solution should be dealy the hrtimer_cancel() in a work
queue, and also a potential solution is move the hrtimer member to the
task_group as the rt sched class and fair sched class did, then we only
need one hrtimer for the whole sched group and free it when freeing the
sched group.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>