RT: add an interrupt latency driver
This is a driver for testing the interrupt latency of a real time
machine, to enable it work on your machine, you just need to implement
the arch specific part in drivers/platform/rt/interrupt_latency.c.

The result with IRQF_NODELAY is around 70us, and the result with a high
prio via "chrt -p prio pid" is also good, about 70us, but the result
with the default prio may miss the interrupt sometime, which means that
the result maybe more than 2ms, but the average result is also good.

Usage:

+ defined UNTHREAD_INTERRUPT in the driver and test it

$ insmod interrupt_latency.ko
$ dmesg | grep mknod
$ mknod /dev/interrupt_latency c 253 0
$ watch /dev/interrupt_latency

+ undefine UNTHREAD_INTERRUPT in the driver to thread the interrupt
handler and use the default prio of the thread.

The usage is the same as above.

+ undefine UNTHREAD_INTERRUPT in the driver to thread the interrupt
handler, but set the

$ insmod interrupt_latency.ko
$ dmesg | grep mknod
$ mknod /dev/interrupt_latency c 253 0
// disable the interrupt and clear the old result
$ echo 0 > /dev/interrupt_latency
// get the pid of the interrupt thread
$ ps -ef | grep irq-interrupt
1443
// change the prio of the interrupt thread
$ chrt -p 99 1443
// reenable the interrupt and restart the tracing
$ echo 1 > /dev/interrupt_latency
// get the result
$ cat /dev/interrupt_latency

Using this driver with the Ftracer may help to find the cause of
interrupt latency: enabling this driver with the Ftracer, when hitting
an indicated big interrupt latency, stop the Ftrace, and get the result
of Ftrace, analyze it, get the cause and fix it.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>