RT: backfire: sendme.c: add context switch latency
This patch records the latency of context switch and file reading.

+ signalhandler(registered in sendme)

	gettimeofday(&tv, NULL);				(3)
	after = tv;

+ sendme
	...
	write("/dev/backfire", sigtest, strlen(sigtest));	(1)
	while (after.tv_sec == 0)				(4)
		;
	gettimeofday(&switchtime, NULL);			(5)

+ backfire

	if (sscanf(buf, "%d", &signo) >= 1) {			(2)
		if (signo > 0 && signo < 32) {
			pid = get_pid(task_pid(current));
			do_gettimeofday(&sendtime);
			kill_pid(pid, signo, 1);
		} else
			printk(KERN_ERR "Invalid signal no. %d\n", signo);
	}

(1) sendme use system call to write a signo into /dev/backfire and
trigger the calling to backfire_write() in the backfire driver.
(2) then the backfire_write get the signo, pid of the current
task(sendme), record the send time and fire a signal to the task.
(3) return from system call, and check the pending signal, execute the
signal handler and record the after time.
(4) try to find the highest prio job and switch to it, now we go back to
sendme.
(5) record the switch time, then we get the context switch latency.

At last, we get the context switch time = switchtime - after.

BTW: the file reading time is also recorded in this patch, and the
default backfire have recored the write time.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>