RT: interrupt latency driver: add dynamic interrupt period support
This patch add dynamic period of interrupt support, users can set the
interrupt period via the user-space interface.

The real period is the user input + the default PERIOD of the interrupt.

A good interval is about 500000 us = 500 ms = 0.5s, no events will be
missed in such an interval.

Herein, the reason for choosing the PERIOD of the timer as 100 is we
want to get a better precision for the precision is 1/tick_rate * count,
the 1/tick_rate is fixed, so, if the count is smaller, we will get a
better precision. and the count is directly related to PERIOD, but we
can not set the PERIOD too small for the system need some time to deal
with the other jobs.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>