RT: interrupt latency driver: Cleanup of the comments
The comments are out-of-date, just cleaned them.

TODO: some of the interfaces are 'confused' currently, for example, we
need to split the set_interval() and record_time() operation out of the
arch_irq_enable() function:

set_interval: set the interval of the interrupt.
record_time: record the interrupt and handling time.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>