RT: interrupt latency driver: lightweight arch_irq_eanble
This patch splits the arch_irq_set_compare() out of the old
arch_irq_enable() and make it more lightweight, the
arch_irq_set_compare() is only needed to be called when we firstly
set/change the compare, so, we put it into irq_enable() instead.

And herein, for the interval < 1000, we can not call msleep() to do
schedule() explicitly for we need a more exact interval indicated by
users, we will only go to schedule() when exit from the interrupt
handler. before, I just guess msleep(0) will not call schedule(), but in
reality, it set the timeout as msec_to_jiffies(0)+1 then called
schedule_timeout(timeout), which will also do schedule().

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>