RT: sched_clock latency driver: enable verbose output
This patch cleaned up the avg calculation and enabled the verbose output
of later statstic, such as Std. calculation, but there is no better
place to print the verbose output, so, just print it into the kernel
buffer then we can dump it out via dmesg.

Since the log buffer of kernel is limited, please use the -s option of
demsg to get more or input a smaller loop.

BTW: all of the printk() are cleaned up with pr_xxx()

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>