Perf: MIPS: Tmp fixup of tools/perf/buitin-record.c
---------
"perf record" works fine on 24K/34K/74K cores. In addition, If you are
seeing the message "Couldn't record kernel reference relocation symbol",
and your kernel symbols only have _stext (without _text), then search
"_text" in builtin-record.c and replace with "_stext".  Here is the
link: http://lkml.org/lkml/2010/1/18/177
---------

Ref: http://www.linux-mips.org/archives/linux-mips/2010-04/msg00179.html

Note: This fixup will break the support of the other architectures, so,
it is only a tmp fix for MIPS, to make it work well on the other
architecture, perhaps we need to find a better solution, such as
introducing a macro _TEXT and define it as _stext for the other
platforms, and define it as _text for MIPS.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>