rt/loongson: add 4 tools
4 different tools are added again for I have added them as submodules
before but forgotten pushing them into the remote git repo.

submodules is not a good idea here, I prefer the directories, so, I
pushed all of the changelogs into the ChangeLog file of them, and
removed all of the .git directory, if you want to get the difference
between them and the mainline versions, you just need to diff ...

So, get the latest same commit them and compare the last changes,
althoug this is not a good idead to remove the history, but anyway, I
don't like submodules, If you really need the git repo, please contact
me, I have backuped them in my own computer.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>