MIPS: Add current_cpu_prid() to optimize the code generation
[ Backport from tiny36 branch of
http://dev.lemote.com/code/linux-loongson-community ]

current_cpu_prid(), cpu_prid_comp(), cpu_prid_imp() and cpu_prid_rev()
are added to simplify/beautify the processord_id related code.

And if current_cpu_prid() is pre-encoded for the specific processor in
cpu-feature-overrides.h, the code generation will be optimized.

cpu_prid_encode() and cpu_prid_encode_copt() are added to encode the
current_cpu_prid(), the former one can be used by most of the processors
whose 'Company Options' part of the prid register is 0 or is not used by
any of the existing codes. Or current_cpu_prid() can be simply assigned
as the value of read_c0_prid(), which can be printed by the
show_cpuinfo() defined in arch/mips/kernel/proc.c.

The size of compressed kernel image(vmlinuz) can be reduced about 0.1M
if current_cpu_prid() is pre-defined as a fixed value in
cpu-feature-overrides.h.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>

Conflicts:

	arch/mips/include/asm/mach-au1x00/au1000.h
	arch/mips/kernel/cpu-probe.c
	arch/mips/loongson/common/env.c

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>